+++ GameServer/Player.cpp
@@ -1,7 +1,7 @@
 #include "Player.h"
 
 const float Player::INITIAL_SPEED = 5;
-const float Player::FIXED_DELTA_TIME = 0.02;
+const float Player::FIXED_DELTA_TIME = (float)0.02;
 #pragma region public
 void Player::move(Move_Req& req)
 {
